VisionSentry-dex-rgb-model-replication/requirements.txt
ultralytics>=8.3.0
opencv-python>=4.8.0
numpy>=1.24.0
PyYAML>=6.0
matplotlib>=3.8.0
pandas>=2.0.0
tqdm>=4.66.0
notebook>=7.0.0
ipykernel>=6.29.0
